add if around rm'ing the files dir

+++ dev/deploy.sh
@@ -131,7 +131,7 @@ ${drush} pm-disable beanstalkd
 
 # Link up the files directory
 drupal_files=$("${web_path}/htdocs/$(${drush} status | sed -ne 's/^ *File directory path *: *\([^ ]*\).*$/\1/p')")
-rm -rf ${drupal_files}
+if [ -d "${drupal_files}" ]; then rm -rf ${drupal_files}; fi
 ln -s /media/nfs/${fqdn} ${drupal_files}
 
 # Sync xhprof webapp directory
